Game "Azukish" v1.0
Created the repository just now, so the commits will start from now on. At the moment there are sprites of the player with a picture, enemies, including UFOs with a picture, meteors large and small without a picture, as well as the main enemies. The player has bullets that he shoots.